refactor(discovery): decompose the traceroute god-file by protocol (WS-D)
Split internal/discovery/traceroute.go (857 LOC) into four role-based
files in the same package (all //go:build !windows) — pure move:

- traceroute.go  shared types, constructors, and cross-protocol helpers
- icmp.go        ICMP probing + TraceICMP/TraceICMPStreaming
- udp.go         UDP probing + TraceUDP
- tcp.go         TCP probing + TraceTCP

All four are now under the 600-line size-gate threshold. Declaration
sets and code-line multisets verified identical to the original.

Related to #187
